# common-platform-customer-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-customer-mgmt
  description: Customer Management Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/customers
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/parties/{partyId}/party-economic-activities/{partyEconomicActivityId}:
    get:
      tags:
        - Party Economic Activities
      summary: Get party economic activity by ID
      description: Retrieve a specific party economic activity associated with a party by its unique identifier
      operationId: getPartyEconomicActivityById
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: partyEconomicActivityId
          in: path
          description: Unique identifier of the party economic activity
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Party economic activity found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PartyEconomicActivityDTO'
        '404':
          description: Party economic activity or party not found
    put:
      tags:
        - Party Economic Activities
      summary: Update party economic activity
      description: Update an existing party economic activity associated with a party
      operationId: updatePartyEconomicActivity
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: partyEconomicActivityId
          in: path
          description: Unique identifier of the party economic activity
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartyEconomicActivityDTO'
        required: true
      responses:
        '200':
          description: Party economic activity successfully updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PartyEconomicActivityDTO'
        '400':
          description: Invalid party economic activity data
        '404':
          description: Party economic activity or party not found
    delete:
      tags:
        - Party Economic Activities
      summary: Delete party economic activity
      description: Delete a party economic activity associated with a party from the system
      operationId: deletePartyEconomicActivity
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: partyEconomicActivityId
          in: path
          description: Unique identifier of the party economic activity
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Party economic activity successfully deleted
        '404':
          description: Party economic activity or party not found
...
  /api/v1/parties/{partyId}/party-economic-activities:
    post:
      tags:
        - Party Economic Activities
      summary: Create party economic activity for a party
      description: Create a new party economic activity associated with a specific party
      operationId: createPartyEconomicActivity
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartyEconomicActivityDTO'
        required: true
      responses:
        '201':
          description: Party economic activity successfully created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PartyEconomicActivityDTO'
        '400':
          description: Invalid party economic activity data
        '404':
          description: Party not found
  /api/v1/parties/{partyId}/party-economic-activities/filter:
    post:
      tags:
        - Party Economic Activities
      summary: Filter party economic activities for a party
      description: Retrieve a paginated list of party economic activities associated with a specific party based on filtering criteria
      operationId: filterPartyEconomicActivities
      parameters:
        - name: partyId
          in: path
          description: Unique identifier of the party
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestPartyEconomicActivityDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved filtered party economic activities
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
        '404':
          description: Party not found
...
components:
  schemas:
...
    PartyEconomicActivityDTO:
      required:
        - economicActivityId
        - partyId
      type: object
      properties:
        partyEconomicActivityId:
          type: string
          format: uuid
          readOnly: true
        partyId:
          type: string
          format: uuid
        economicActivityId:
          type: string
          format: uuid
        annualTurnover:
          minimum: 0
          exclusiveMinimum: false
          type: number
        currencyCode:
          maxLength: 3
          minLength: 0
          pattern: ^[A-Z]{3}$
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        isPrimary:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestPartyEconomicActivityDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/PartyEconomicActivityDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for party economic activities
...